+++ public/manifest.webmanifest
@@ -1,5 +1,5 @@
 {
-  "name": "T&P Cell | GEC, Siwan",
+  "name": "Training and Placement Cell, GEC Siwan",
   "short_name": "T&P Cell",
   "start_url": ".",
   "display": "standalone",
