+++ src/styles/index.css
@@ -13,6 +13,12 @@
 html {
   font-size: 14px;
 }
+main{
+  background-image: url('/images/blur_bg.png');
+  background-size: cover;
+  background-repeat: no-repeat;
+  background-position: center;
+}
 body {
   margin: auto;
   max-width: 100%;
